% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{Find the oldest person}
\usage{
oldest(x)
}
\arguments{
\item{x}{A parsed Williams faculty dataset to be searched through}
}
\value{
The name of the oldest faculty member
}
\description{
Takes in a catalog dataset full of Williams faculty members and finds the oldest person
}

