% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youngest.R
\name{youngest}
\alias{youngest}
\title{Find the youngest person}
\usage{
youngest(x)
}
\arguments{
\item{x}{A parsed Williams faculty dataset to be searched through}
}
\value{
The name of the youngest faculty member
}
\description{
Takes in a catalog dataset full of Williams faculty members and finds the youngest person
}

